References

Speculator.speculate — Function

speculate(predicate, value; parameters...)
speculate(value; parameters...)

Generate a compilation a workload.

See also install_speculator.

Tip

Use this in a package to reduce latency for its users.

Note

Speculation only runs when called during precompilation or an interactive session, or when writing precompilation directives to a file.

Parameters

predicate = Returns(true): This must accept the signature predicate(::Module, ::Symbol)::Bool. Returning true specifies to search getproperty(::Module, ::Symbol), whereas returning false specifies to ignore the value. This is called when searching the names of a Module if the given module and name satisfy isdefined and !isdeprecated. The default predicate Returns(true) will search everything possible, up to the generic limit, whereas the predicate Returns(false) will not generate any methods. Some useful predicates include Base.isexported, Base.ispublic, checking properties of the value itself, and a combination thereof.

value: When given a Module, speculate will recursively search its contents using names(::Module; all = true), for each name that is not deprecated, is not an external module, and satisifes the predicate. For other values, each of their generic methods are searched for corresponding compilable signatures.

Keyword parameters

background::Bool = false: Specifies whether to precompile on a thread in the :default pool. The number of available threads can be determined using Threads.nthreads(:default).

dry::Bool = false: Specifies whether to run precompile on generated method signatures. This is useful for testing workloads with verbosity = debug ∪ review. Methods that are known to be specialized are skipped. Note that dry must be false to save the workload to a file with the path parameter.

limit::Int = 1: Specifies the maximum number of compilable methods that are generated from a generic method. Values less than 1 will throw an error. Otherwise, method signatures will be generated from the Cartesian product each parameter type. Types marked with @nospecialize are used directly. Otherwise, compilable types are obtained from the subtypes of DataType and Union. This prevents spending too much time precompiling a single generic method.

path::String = "": Saves a workload by writing each successful precompilation directive to a file if the path is not empty and it is not a dry run. Note that these directives may require loading additional modules to run.

verbosity::Verbosity = warn: Specifies what logging statements to show. If this function is used in a package as a precompilation workload, this should be set to silent or warn. See also Verbosity.

Examples

julia> module Showcase
           export g, h

           f() = nothing
           g(::Int) = nothing
           h(::Union{String, Symbol}) = nothing
       end;

julia> speculate(Showcase; verbosity = debug)
[ Info: Compiled `Main.Showcase.g(::Int)`
[ Info: Compiled `Main.Showcase.f()`

julia> speculate(Base.isexported, Showcase; verbosity = debug)
[ Info: Skipped `Main.Showcase.g(::Int)`

julia> speculate(Showcase.h; limit = 2, verbosity = debug)
[ Info: Compiled `Main.Showcase.h(::String)`
[ Info: Compiled `Main.Showcase.h(::Symbol)`

Input Speculator

Speculator.install_speculator — Function

install_speculator(predicate = Returns{Bool}(true); background::Bool = true, parameters...)

Install an input speculator that calls speculate(predicate, value; background, parameters...) on each input value in the REPL.

Subsequent calls to this function may be used to replace the hook. The hook may be removed using uninstall_speculator. This function has no effect in non-interactive sessions.

See also speculate.

Tip

Use this in a startup.jl file to reduce latency in the REPL. Since it relies on the REPL being initialized, it should be placed at the end of the file.

julia> install_speculator(; limit = 2, verbosity = debug)
[ Info The input speculator has been installed into the REPL

julia> f() = nothing;
[ Info: Compiled `Main.Example.f()`

julia> g(::Union{String, Symbol}) = nothing;
[ Info: Compiled `Main.Example.g(::String)`
[ Info: Compiled `Main.Example.g(::Symbol)`

Speculator.uninstall_speculator — Function

uninstall_speculator()

Uninstall the hook that may have previously been installed by install_speculator.

julia> uninstall_speculator()
[ Info: The input speculator has been uninstalled from the REPL

All Modules

Speculator.AllModules — Type

AllModules

A singleton type whose only value is all_modules.

Interface

show(::IO, ::AllModules)

Examples

julia> AllModules
AllModules

Speculator.all_modules — Constant

all_modules::AllModules

The singleton constant of AllModules used with speculate to generate a compilation workload using all loaded modules.

Examples

julia> all_modules
all_modules::AllModules

Verbosities

Speculator.Verbosity — Type

Verbosity

A flag that determine what logging statements are shown during speculate.

This is modelled as a set, where silent is the empty set. The non-empty component flags are debug, review, and warn.

Interface

This type implements part of the AbstractSet interface.

intersect(::Verbosity, ::Verbosity...)

isdisjoint(::Verbosity, ::Verbosity)

isempty(::Verbosity)

issetequal(::Verbosity, ::Verbosity)

issubset(::Verbosity, ::Verbosity)

setdiff(::Verbosity, ::Verboosity...)

show(::IO, ::Verbosity)

union(::Verbosity, ::Verbosity...)

Examples

julia> silent
silent::Verbosity

julia> debug ∪ review
(debug ∪ review)::Verbosity

julia> debug ⊆ debug ∪ review
true

julia> debug ⊆ warn
false

Speculator.debug — Constant

debug::Verbosity

A flag of Verbosity which specifies that speculate will show each successful call to precompile.

Examples

julia> debug
debug::Verbosity

Speculator.review — Constant

review::Verbosity

A flag of Verbosity which specifies that speculate will show a summary of the number of methods generated, the number of generic methods found, and the duration. If dry = false, this also shows the number of generated methods that were compiled, skipped, and warned.

Examples

julia> debug
debug::Verbosity

Speculator.silent — Constant

silent::Verbosity

A flag of Verbosity which specifies that speculate will not show any logging statements.

Examples

julia> silent
silent::Verbosity

Speculator.warn — Constant

warn::Verbosity

A flag of Verbosity which specifies that speculate will show warnings for failed calls to precompile.

Examples

julia> warn
warn::Verbosity