References

Speculator.SpeculationBenchmark — Type

SpeculationBenchmark <: Any
SpeculationBenchmark(predicate = Returns(true), value;
    limit::Integer = 1, trials::Integer = 8
)

Benchmark the compilation time in the workload generated by speculate(predicate, value; samples, limit).

This runs a compilation workload for each of the trials. Each trial occurs in a new process so that compiled methods are not cached across trials. The trials run the workload with dry = true to compile methods in Speculator.jl, then dry = false to measure the runtime within Speculator.jl and calls to precompile, and finally dry = false to measure the runtime within Speculator.jl. The result of a trial, an estimate of the runtime of calls to precompile in the workload, is the difference between the second and third runs.

See also speculate_repl and speculate.

Tip

Initializing a temporary project and running precompilation workloads may take a substantial amount of time to complete. It is recommended to select an appropriate workload using speculate(; dry = true, verbosity = debug | review) before creating a benchmark.

Interface

This type implements the iteration interface and part of the indexing interface.

eltype(::Type{<:SpeculationBenchmark})

firstindex(::SpeculationBenchmark)

getindex(::SpeculationBenchmark, ::Integer)

iterate(::SpeculationBenchmark, ::Integer)

iterate(::SpeculationBenchmark)

lastindex(::SpeculationBenchmark)

length(::SpeculationBenchmark)

show(::IO, ::MIME"text/plain", ::SpeculationBenchmark)

Speculator.speculate_repl — Function

speculate_repl(predicate = Returns{Bool}(true);
    install::Bool = true, background::Bool = true,
parameters...)

Install a hook that calls speculate(predicate, value; background, parameters...) on each input value in the REPL.

Subsequent calls to this function may be used to replace the hook. The hook may be removed using speculate_repl(; install = false). This function has no effect in non-interactive sessions.

See also SpeculationBenchmark and speculate.

Tip

Use this in a startup.jl file to reduce latency in the REPL.

julia> speculate_repl(; limit = 2, verbosity = debug)
[ Info: The REPL will call `speculate` with each input

julia> f() = nothing;
[ Info: Compiled `Main.Example.f()`

julia> g(::Union{String, Symbol}) = nothing;
[ Info: Compiled `Main.Example.g(::String)`
[ Info: Compiled `Main.Example.g(::Symbol)`

julia> speculate_repl(; install = false)
[ Info: The REPL will not call `speculate` with each input

Speculator.speculate — Function

speculate(predicate, value; parameters)
speculate(value; parameters...)

Generate a compilation a workload.

See also SpeculationBenchmark and speculate.

Tip

Use this in a package to reduce latency for its users.

Note

This function only runs when called during precompilation or an interactive session, or when writing precompilation directives to a file.

Parameters

predicate = Returns(true): This must have the signature predicate(::Module, ::Symbol)::Bool, and is checked for each name given by names(::Module; all = true). Returning true specifies to search getproperty(::Module, ::Symbol), whereas returning false specifies to ignore the value. The default predicate Returns{Bool}(true) will search every possible method, up to its generic limit, whereas the predicate Returns(false) will only search for methods of values passed directly to speculate. This new value can also be accessed in the same manner within the predicate. Some useful predicates include Base.isexported, Base.ispublic, checking properties of the value itself, and a combination thereof.

value: When given a Module, speculate will recursively search its contents using names(::Module; all = true). For each other value, each of their generic methods are searched for corresponding compilable signatures.

Keyword parameters

background::Bool = false: Specifies whether to precompile on a thread in the :default pool. The number of available threads can be determined using Threads.nthreads(:default). In an interactive session with debug in the verbosity, a call to sleep(0.01) is used to keep the REPL prompt active.

dry::Bool = false: Specifies whether to run precompile on generated method signatures. This is useful for testing workloads with verbosity = debug. Methods that are known to be specialized are skipped. Note that dry must be false to save the workload to a file with the path parameter.

limit::Integer = 1: Specifies the maximum number of compilable methods that are generated from a generic method. Values less than 1 will throw an error. Otherwise, method signatures will be generated from the Cartesian product each parameter type. Concrete types and abstract types marked with @nospecialize are used directly. Otherwise, compilable types are obtained from the subtypes of DataType and Union. This prevents spending too much time precompiling a single generic method.

path::String = "": Writes each successful precompilation directive to a file if the path is not empty and it is not a dry run. Note that these directives may require loading additional modules to run.

verbosity::Verbosity = warn: Specifies what logging statements to show. If this function is used as a precompilation workload, this should be set to silent or warn. See also Verbosity.

Examples

julia> module Showcase
           export g, h

           f() = nothing
           g(::Int) = nothing
           h(::Union{String, Symbol}) = nothing
       end;

julia> speculate(Showcase; verbosity = debug)
[ Info: Compiled `Main.Showcase.g(::Int)`
[ Info: Compiled `Main.Showcase.f()`

julia> speculate(Base.isexported, Showcase; verbosity = debug)
[ Info: Skipped `Main.Showcase.g(::Int)`

julia> speculate(Showcase.h; limit = 2, verbosity = debug)
[ Info: Compiled `Main.Showcase.h(::String)`
[ Info: Compiled `Main.Showcase.h(::Symbol)`

Verbosity

Speculator.Verbosity — Type

Verbosity

A flag that determine what logging statements are shown during speculate.

The component flags are silent, debug, review, and warn.

Interface

|(::Verbosity, ::Verbosity)

Combine the verbosities such that each satisfies issubset with the returned verbosity.

issubset(::Verbosity, ::Verbosity)

Check whether each flag of the first verbosity is a component of the second verbosity.

show(::IO, ::Verbosity)

Examples

julia> silent
silent::Verbosity

julia> debug | review
(debug | review)::Verbosity

julia> debug ⊆ debug | review
true

julia> debug ⊆ warn
false

Speculator.debug — Constant

debug::Verbosity

A flag of Verbosity which specifies that speculate will show each successful call to precompile.

Examples

julia> debug
debug::Verbosity

Speculator.review — Constant

review::Verbosity

A flag of Verbosity which specifies that speculate will show the total number of values that have been speculated.

Examples

julia> debug
debug::Verbosity

Speculator.silent — Constant

silent::Verbosity

A flag of Verbosity which specifies that speculate will not show any logging statements.

Examples

julia> silent
silent::Verbosity

Speculator.warn — Constant

warn::Verbosity

A flag of Verbosity which specifies that speculate will show warnings for failed calls to precompile.

Examples

julia> warn
warn::Verbosity